WIP SceneIter
warning fixes

+++ src/nest/NEST_neighbor_test.cc
@@ -428,7 +428,7 @@ void generic_test_neighbors(
 	typedef typename NestType::IndexType Index;
 	typedef std::vector<typename NestType::IndexType> IndexVec;
 	typedef std::set<typename NestType::IndexType> IndexSet;
-	boost::random::mt19937 rng(time(0));
+	boost::random::mt19937 rng((unsigned int)time(0));
 	boost::uniform_real<> uniform;
 	typename NestType::ValueType randpt;
 	for(size_t r = 0; r <= 9; ++r){
